% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_cor2.R
\name{cor2}
\alias{cor2}
\title{A substitute for cor() supporting sparse matrices}
\usage{
cor2(X, Y = NULL, use = "everything", center = TRUE)
}
\arguments{
\item{X}{base matrix}

\item{Y}{base matrix, defaults to X}

\item{use}{missing value handling: "everything" (default) or "na.or.complete"}

\item{center}{logical: center columns (default TRUE)}
}
\value{
p x q matrix of correlations
}
\description{
A substitute for cor() supporting sparse matrices
}
