% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_crate.R
\name{crate}
\alias{crate}
\title{Create Self-Contained Functions}
\usage{
crate(
  .fn,
  ...,
  .parent_env = baseenv(),
  .error_arg = ".fn",
  .error_call = environment()
)
}
\arguments{
\item{.fn}{A fresh formula or function. "Fresh" here means that
they should be declared in the call to \code{crate()}. See examples if
you need to crate a function that is already defined. Formulas
are converted to purrr-like lambda functions using
\code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{...}{Objects to include in the crate}

\item{.parent_env}{The default of \code{baseenv()} ensures that the evaluation
environment of the crate is isolated from the search path. Specifying
another environment such as the global environment allows this condition to
be relaxed (but at the expense of no longer being able to rely on a local
run giving the same results as one in a different process).}

\item{.error_arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{.error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A crated function with all dependencies included
}
\description{
Creates a self-contained function that contains all necessary dependencies.
This is useful for parallel processing where functions need to be serialized
and sent to different workers.
}
\examples{
\dontrun{
# Create a function with dependencies
data <- 1:10
crated_fn <- crate(function(x) x + mean(data), data = data)
crated_fn(5)
}

}
\seealso{
\code{\link[carrier:crate]{carrier::crate()}}
}
