% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_GetPythonPath.R
\name{GetPythonPath}
\alias{GetPythonPath}
\title{Get Python Executable Path}
\usage{
GetPythonPath(path)
}
\arguments{
\item{path}{Character string specifying the path to the environment where Python might be located.}
}
\value{
Character string with the normalized path to the Python executable if found,
otherwise \code{NA_character_}.
}
\description{
This function attempts to find the Python executable within a given environment path.
It checks different candidate paths based on the operating system.
}
