% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigBridgeRUtils-package.R
\docType{package}
\name{SigBridgeRUtils-package}
\alias{SigBridgeRUtils-package}
\alias{SigBridgeRUtils}
\title{Some Utilities & Base Supports for 'SigBridgeR'}
\description{
Provides fundamental function support for 'SigBridgeR' and its single-cell phenotypic screening algorithm, including optional functions.
}
\section{License}{

GPL (>= 3)
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/SigBridgeRUtils}
  \item Report bugs at \url{https://github.com/WangLabCSU/SigBridgeRUtils/issues}
}

}
\author{
\strong{Maintainer}: Yuxi Yang \email{15364051195@163.com} (\href{https://orcid.org/0009-0006-1329-1224}{ORCID})

Authors:
\itemize{
  \item Yuxi Yang \email{15364051195@163.com} (\href{https://orcid.org/0009-0006-1329-1224}{ORCID})
}

}
\keyword{internal}
