% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_caller_cli.R
\name{GetCallerInfo}
\alias{GetCallerInfo}
\title{Get Caller Name}
\usage{
GetCallerInfo(offset = 2)
}
\arguments{
\item{offset}{Integer. The number of stack frames to go back.
Defaults to 2 (skipping the GetCallerInfo frame and the Wrapper frame
to find the User's function).}
}
\value{
Character string. E.g., "my_function()" or "global".
}
\description{
Retrieves the name of the function that called the current execution context.
If called from the global environment, returns "global".
}
\examples{
\dontrun{
f <- function() { GetCallerInfo() }
f() # Returns "f()"
}

}
