% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26_GetFuncArgs.R
\name{GetFuncArgs}
\alias{GetFuncArgs}
\title{Get Function Arguments from Calling Context}
\usage{
GetFuncArgs(
  exclude = NULL,
  name_only = FALSE,
  func = rlang::caller_fn(),
  call = rlang::caller_call(),
  dots_expand = TRUE,
  envir = rlang::caller_env()
)
}
\arguments{
\item{exclude}{A character or interger vector of argument names to exclude from the result. Default is \code{NULL}.}

\item{name_only}{Logical. If \code{TRUE}, returns only argument names. If \code{FALSE}, returns the full argument list. Default is \code{FALSE}.}

\item{func}{which function to seek}

\item{call}{The call expression to extract arguments from. Default is \code{rlang::caller_call()}.}

\item{dots_expand}{Logical. Whether to expand \code{...} arguments. Default is \code{TRUE}.}

\item{envir}{The environment in which to evaluate the call. Default is \code{rlang::caller_env()}.}
}
\value{
If \code{name_only} is \code{TRUE}, a character vector of argument names. Otherwise, a named list of arguments.
}
\description{
Retrieves the arguments passed to the calling function, optionally filtering by name or returning only argument names.
}
\examples{
\dontrun{
inner_func <- function(a, b, ...) {
  GetFuncArgs(exclude = "b", name_only = TRUE)
}
inner_func(1, 2, 3, 4)  # Returns c("a", "...")
}

}
